Pack integer arguments if they fit in 32bit field

+++ syscall_args.hpp
@@ -6,6 +6,6 @@
 class ManagedBuffer;
 
 void writeSyscallArgumentInt(ManagedBuffer &managed_buffer, int64_t arg);
-void writeSyscallArgumentData(ManagedBuffer &managed_buffer, int64_t arg, int32_t size);
+void writeSyscallArgumentData(ManagedBuffer &managed_buffer, int64_t arg, int64_t size);
 
 #endif // SYSCALL_ARGS_HPP
